move run to wl_handlers
